use args instead of deps in default deriv impl
The `WithDerivative` evaluable defines a derivative of a wrapped array to some
target. The target is not included in the `Evaluable.dependencies`, but is
included in the `Array.arguments`. The current default implementation of
`Array._derivative` incorrectly returns zeros if the derivative target is not
found in `self.dependencies`. This patch fixes the bug by testing for existence
in `self.arguments`.