replace session, eval_withtimes by eval_profiling
This patch removes eval_withtimes, which was used only by session, and session,
which was used only by eval_sparse. Instead, a new method eval_profiling is
introduced that is called directly by eval_sparse, and which returns a graph
node that eval_sparse can use to draw an asciitree or graphviz diagram. While
eval_sparse is perhaps not the most logical place for this code, it's ok for
now, and a new auxiliary function is easily introduced once the best structure
for it becomes clear (we can be fairy certain that we will not want to walk
back the removal of session, in any case).